+++ src/components/semiotic-data.ts
@@ -1,0 +1,5 @@
+/**
+ * Semiotic data transform utilities.
+ * Import from "semiotic/data"
+ */
+export { bin, rollup, groupBy, pivot } from "./data/transforms"